requirements.txt
aiofiles==23.1.0
autocommand==2.2.2
Babel==2.8.1
blinker==1.5
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
coloredlogs==15.0.1
docopt==0.6.2
flatbuffers==23.3.3
gruut==1.3.1
gruut-ipa==0.10.1
gruut-lang-de==1.3.1
gruut-lang-es==1.3.1
gruut-lang-fr==1.3.1
gruut-lang-it==1.3.1
gruut-lang-nl==1.3.1
gruut-lang-ru==1.3.1
gruut-lang-sv==1.3.1
gruut-lang-sw==1.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
humanfriendly==10.0
hypercorn==0.14.3
hyperframe==6.0.1
inflect==6.0.2
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
Jinja2==3.1.2
jsonlines==1.2.0
jsons==1.6.3
larynx==0.5.0
loguru==0.6.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
num2words==0.6.0
numpy==1.24.2
onnxruntime==1.14.1
packaging==23.0
portend==3.1.0
priority==2.0.0
protobuf==4.22.1
psutil==5.8.0
pydantic==1.10.6
pyscard==2.0.5
pyserial==3.5
python-crfsuite==0.9.9
python-pidfile==3.0.0
pyttsx3==2.90
pytz==2022.7.1
PyYAML==6.0
Quart==0.11.5
Quart-CORS==0.3.0
six==1.16.0
swagger-ui-py==0.2.1
sympy==1.11.1
tempora==5.2.1
toml==0.10.2
tqdm==4.48.2
typing_extensions==4.5.0
typish==1.9.3
Werkzeug==2.2.3
wsproto==1.2.0
zc.lockfile==3.0.post1
